requirements.txt
beautifulsoup4==4.12.3
lxml==5.1.0
fake-useragent==1.4.0
aiohttp==3.9.1
asyncio==3.4.3
SQLAlchemy==2.0.25
alembic==1.13.1
fastapi==0.109.0
python-dotenv==1.0.0
fastapi-users-db-sqlalchemy==6.0.1
celery==5.3.6
asyncpg==0.29.0
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
redis==5.0.1
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.26.0
gunicorn==21.2.0